Update UserDAOImpl.java
Bug Fix: Changed the 'whereClause' in the updateUser() method to reference the rowid of the user. This will allow the email(username) to be updated from the 'userMangementPage'. Otherwise, the statement will fail to find the reference to the 'new email' in the users table as it does not yet exist.